refactor: ヘルパー関数をutilsディレクトリに移動
